fix(rendering): correct CPU camera orbit/pan and timer state
The orbit eye had two contradictory derivations: viewMatrix() placed the
camera at target + distance*sin(pitch) on Y while updateFreeFromOrbit()
used -distance*sin(pitch). Since forward is (-sin(y)cos(p), sin(p),
-cos(y)cos(p)), the eye is target - distance*forward, so the offset is
negative; the positive variant viewed the scene from the mirrored
elevation and switching Orbit->Free teleported the camera. Both now read
one orbitEye() helper, and the mode round trip is view-matrix stable.

pan() and move() each hand-rolled a roll-free basis, so drag axes stopped
matching what the user saw once the gizmo rolled the view. Both consume a
single cameraBasis() that rotates up/right about forward by roll.

Rotation feedback was truncated to int degrees, so sub-degree orbit drags
reached the gizmo as zero. cameraRotated, cameraRotationChanged and
setCameraRotation are float end to end; MainWindow wiring is unchanged
because the float->float function-pointer connects still match.

updatePhysics() divided QElapsedTimer::restart() by 1000, but that call
returns nanoseconds, so the integrator ran ~1000x fast on top of a first
tick that measured construction plus first paint. The conversion is now
ns -> seconds, and clampFrameDeltaSeconds() replaces only the first frame
with one nominal kPhysicsTickSeconds tick. The policy lives in the library
so it is headless-testable.

NavigationGizmo set ClosedHandCursor on press and never reset it, leaving
a clenched fist for the widget's lifetime; a dragging_ flag plus
mouseReleaseEvent/enterEvent/leaveEvent restore Qt::OpenHandCursor.

tests/camera_basis_contract.cpp asserts the basis, eye and view matrix
against a hand-derived oracle, the negative-Y convention as a strict sign
fact and an invariant, roll-aware pan and free-flight axes, mode round
trip continuity, rotate sensitivities with the pitch clamp, and the
first-frame tick policy for zero/absurd/NaN elapsed. 172 checks, stable
across OMP_NUM_THREADS 1/2/4/8, and it fails against four re-broken
camera variants. Registered as the 31st CPU test in both registries, with
anti-synthetic gate 10 requiring real call shapes and the real Camera TU
inside azu_test_core (verified to fail closed).

CPU lane only: no CUDA/HIP configured, compiled or runtime-tested, and the
protected build-cpu/ and build-cuda/ trees are untouched.

+++ include/gui/ControlPanel.h
@@ -37,14 +37,14 @@ signals:
     void volumeCageToggled(bool visible);
     void threadsChanged(int n);
     void hyperparamsApplyClicked();
-    void cameraRotationChanged(int pitch, int yaw, int roll);
+    void cameraRotationChanged(float pitch, float yaw, float roll);
 
 public slots:
     void onPipelineStarted();
     void onPipelineStopped();
     void setExportEnabled(bool enabled);
     void setBusy(bool busy);
-    void setCameraRotation(int pitch, int yaw, int roll);
+    void setCameraRotation(float pitch, float yaw, float roll);
     void onPresetChanged(int index);
 
 private:
